+++ OneBlink.SDK/models/FormSubmissionEventConfigurationMicrosoftResource.cs
@@ -1,0 +1,14 @@
+namespace OneBlink.SDK.Model
+{
+  public class FormSubmissionEventConfigurationMicrosoftResource
+  {
+    public string id
+    {
+      get; set;
+    }
+    public string displayName
+    {
+      get; set;
+    }
+  }
+}