+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 ### Added
 
 - `isAIBuilderSupported` to `Form`
+- properties to to `FormEvent` for excel add row event
 - `hideApprovalDenyButton` to `FormApprovalFlowStepBase`
+
 
 ## [11.1.1] - 2025-09-15